Pi data parsed, analysis script updated

+++ raspi/localization/scan_result.h
@@ -12,7 +12,7 @@ using namespace std;
 
 struct Result {
   string name;
-  int signal;
+  double signal;
 
   template<class Archive>
   void serialize(Archive & archive) {
